# elasticSearch/trabalho-pratico-elasticsearch/codigo-modificado/src/main/resources/api.yml
openapi: '3.0.0'
info:
  version: '2.0.0'
  title: 'Wikipedia Search API'
  description: |
    API RESTful para busca de documentos Wikipedia via Elasticsearch.
    Suporta busca full-text com fuzzy matching, highlight de termos,
    paginação com total de resultados e sugestões de correção ortográfica.
  contact:
    name: 'Elasticsearch Example'
  license:
    name: 'MIT'

servers:
  - url: 'http://localhost:8080/v1'
    description: 'Servidor local de desenvolvimento'

tags:
  - name: search
    description: 'Operações de busca full-text no índice Wikipedia'
  - name: suggest
    description: 'Sugestões de correção ortográfica e autocomplete'

paths:
  /search:
    get:
      summary: 'Realiza busca full-text no índice Wikipedia'
      description: |
        Executa uma query booleana com must + should (phrase boost) contra
        os campos `content` e `title` do índice Wikipedia. Suporta fuzzy
        matching automático (fuzziness: auto) e highlight de termos encontrados.
        Retorna cabeçalhos HTTP com metadados de paginação.
      operationId: search
      tags:
        - search
      parameters:
        - name: query
          in: query
          required: true
          description: 'Termo ou frase a ser buscado. Suporta erros de digitação via fuzzy matching.'
          schema:
            type: string
            minLength: 1
            example: 'square root'
        - name: page
          in: query
          required: false
          description: 'Número da página de resultados (1-indexed)'
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: size
          in: query
          required: false
          description: 'Quantidade de resultados por página'
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
            example: 10
      responses:
        200:
          description: 'Lista de resultados encontrados (pode ser vazia)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        400:
          description: 'Parâmetro query ausente ou inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: 'Erro interno do servidor ou falha na conexão com Elasticsearch'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /suggest:
    get:
      summary: 'Sugestões de correção ortográfica para um termo'
      description: |
        Utiliza o mecanismo de Term Suggest do Elasticsearch para corrigir
        erros de digitação e sugerir termos alternativos baseados no vocabulário
        do índice Wikipedia. Útil para o recurso "Você quis dizer?" na UI.
      operationId: suggest
      tags:
        - suggest
      parameters:
        - name: query
          in: query
          required: true
          description: 'Texto com possíveis erros ortográficos para o qual se deseja sugestões'
          schema:
            type: string
            minLength: 1
            example: 'Kolmogorv equatons'
        - name: size
          in: query
          required: false
          description: 'Número máximo de sugestões por termo'
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
            example: 3
      responses:
        200:
          description: 'Lista de termos sugeridos (pode ser vazia se a query já estiver correta)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
              example:
                original: 'Kolmogorv equatons'
                suggestions:
                  - 'kolmogorov equations'
                  - 'kolmogorov equation'
                hasSuggestion: true
        500:
          description: 'Erro interno do servidor'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    SearchResponse:
      type: object
      description: 'Objeto contendo os resultados da busca e os metadados de paginação'
      properties:
        results:
          type: array
          description: 'Lista de resultados encontrados'
          items:
            $ref: '#/components/schemas/Result'
        totalCount:
          type: integer
          description: 'Número total de documentos que correspondem à query'
          example: 137
        totalPages:
          type: integer
          description: 'Número total de páginas disponíveis (ceil(total / size))'
          example: 14
        currentPage:
          type: integer
          description: 'Página atual retornada'
          example: 1
        pageSize:
          type: integer
          description: 'Quantidade de resultados por página'
          example: 10

    Result:
      type: object
      description: 'Um resultado de busca do índice Wikipedia'
      properties:
        title:
          type: string
          description: 'Título do artigo Wikipedia'
          example: 'Square root'
        url:
          type: string
          description: 'URL original do artigo Wikipedia'
          example: 'https://en.wikipedia.org/wiki/Square_root'
        abs:
          type: string
          description: 'Resumo do conteúdo do artigo (abstract), com termos buscados destacados em <strong>tags</strong>'
          example: 'In mathematics, a <strong>square root</strong> of a number x...'

    SuggestResponse:
      type: object
      description: 'Resposta do mecanismo de sugestão'
      properties:
        original:
          type: string
          description: 'Texto original enviado pelo usuário'
          example: 'Kolmogorv equatons'
        suggestions:
          type: array
          description: 'Lista de sugestões de correção'
          items:
            type: string
          example:
            - 'kolmogorov equations'
            - 'kolmogorov equation'
        hasSuggestion:
          type: boolean
          description: 'Indica se foram encontradas sugestões de correção'
          example: true

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: 'Mensagem de erro legível'
          example: 'O parâmetro query é obrigatório'
        code:
          type: string
          description: 'Código de erro interno'
          example: 'MISSING_QUERY'